pyproject.toml
[project]
name = "browser-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browserbase>=1.4.0",
    "litellm>=1.73.2",
    "loguru>=0.7.3",
    "mcp>=1.10.1",
    "playwright>=1.53.0",
    "pydantic-ai>=0.3.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=13.7.0",
    "stagehand>=0.0.7",
    "uvloop>=0.21.0",
]
